ENH: Addition of an external project to download and compile FLTK automatically.
An option to apply a patch to FLTK has also been added. Some older compiler need a patch to be applied to FLTK to be able to compile it, others do not. Try without the patch and if you run into problems, remove FLTK files that may already have been downloaded and compiled in your build tree, change the value of this option and try to compile FLTK again.