STYLE: clean source code by removing commented out code BUG: old fluid registration and reorientation bug
If AutoSeg was started with the GUI, the tools were detected on the system, and the old fluid registration and reorientation blocks were deactivated if some of the tools necessary for these operations were missing. If AutoSeg was started without the GUI, it was not verifying if the reorientation software and the old fluid registration software were available on the system.