ENH: Update ANTs and ITK revisions
The previous revision of ANTs used was not installing all the necessary libraries in the install directory. The build library had to be given to speficy the ants libraries directory for BRAINSTools to build. This problem has been resolved in the newest revision of ANTs, but this revision requires a more recent version of ITK. Therefore the revision of ITK was also updated.